Fix QRcode String Reader
