--- なにがでるかな/Assets/Scripts/Battle/QRReader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 20d05eacc965527438adf5002afd313a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
