+++ なにがでるかな/Assets/Scripts/Battle/System.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 210b8691ad3a6c94691f50ff082dadd1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
